#300: Moved character-related components to their own folder
 The folder is called character-display-components.
