+++ src/app/ffbe/character-display-components/caracteristiques-ex-display/caracteristiques-ex-display.component.html
