+++ src/app/ffbe/character-display-components/caracteristiques-ex-display/caracteristiques-ex-display.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {FfbeUtils} from '../utils/ffbe-utils';
-import {UniteCarac} from '../model/unite-carac.model';
+import {FfbeUtils} from '../../utils/ffbe-utils';
+import {UniteCarac} from '../../model/unite-carac.model';
 
 @Component({
   selector: 'app-caracteristiques-ex-display',
