+++ src/app/ffbe/character-display-components/character-display/character-display.component.css
