+++ src/app/ffbe/character-display-components/character-equipments-display/character-equipments-display.component.html
