+++ src/app/ffbe/ffbe.module.ts
@@ -4,7 +4,7 @@ import {CharactersComponent} from './characters/characters.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {CharactersService} from './services/characters.service';
 import {DataMiningClientService} from './services/data-mining-client.service';
-import {CharacterDisplayComponent} from './character-display/character-display.component';
+import {CharacterDisplayComponent} from './character-display-components/character-display/character-display.component';
 import {BrowserModule} from '@angular/platform-browser';
 import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
@@ -17,13 +17,13 @@ import {MatSelectModule} from '@angular/material/select';
 import {MatTableModule} from '@angular/material/table';
 import {LimitBurstsService} from './services/limit-bursts.service';
 import {SkillsService} from './services/skills.service';
-import {UnitDisplayComponent} from './unit-display/unit-display.component';
+import {UnitDisplayComponent} from './character-display-components/unit-display/unit-display.component';
 import {SkillsDisplayComponent} from './skills-display/skills-display.component';
 import {FfchClientService} from './services/ffch-client.service';
 import {SkillsComponent} from './skills/skills.component';
 import {SkillDisplayComponent} from './skill-display/skill-display.component';
-import {CharacterEquipmentsDisplayComponent} from './character-equipments-display/character-equipments-display.component';
-import {UnitAwakeningMaterialsDisplayComponent} from './unit-awakening-materials-display/unit-awakening-materials-display.component';
+import {CharacterEquipmentsDisplayComponent} from './character-display-components/character-equipments-display/character-equipments-display.component';
+import {UnitAwakeningMaterialsDisplayComponent} from './character-display-components/unit-awakening-materials-display/unit-awakening-materials-display.component';
 import {EnhancementsComponent} from './enhancements/enhancements.component';
 import {EnhancementsService} from './services/enhancements.service';
 import {EnhancementsDisplayComponent} from './enhancements-display/enhancements-display.component';
@@ -51,7 +51,7 @@ import {ObjetAlterationsEtatDisplayComponent} from './objets-display-components/
 import {LatentSkillsComponent} from './latent-skills/latent-skills.component';
 import {LatentSkillsService} from './services/latent-skills.service';
 import {CaracteristiquesContainerDisplayComponent} from './caracteristiques-container-display/caracteristiques-container-display.component';
-import {CaracteristiquesEXDisplayComponent} from './caracteristiques-ex-display/caracteristiques-ex-display.component';
+import {CaracteristiquesEXDisplayComponent} from './character-display-components/caracteristiques-ex-display/caracteristiques-ex-display.component';
 import {ObjetTueursPhysiquesMagiquesDisplayComponent} from './objets-display-components/objet-tueurs-physiques-magiques-display/objet-tueurs-physiques-magiques-display.component';
 import {ObjetTueursDisplayComponent} from './objets-display-components/objet-tueurs-display/objet-tueurs-display.component';
 import {ObjetsComparingContainerDisplayComponent} from './objets-display-components/objets-comparing-container-display/objets-comparing-container-display.component';
